Add proxy support to HibpDownloader and SyncHashesFromHibp
The `HibpDownloader` class now supports an optional proxy server. A new constructor overload accepts a `Uri` object for the proxy. The `InitializeHttpClient` method configures the `HttpClientHandler` to use the proxy if provided.

The `SyncHashesFromHibp` class has a new `ProxyAddress` parameter. This allows users to specify a proxy server address. The `BeginProcessing` method validates and uses this proxy address when creating an instance of `HibpDownloader`.